+++ TelegramBotAPI-core/src/commonTest/kotlin/com/github/insanusmokrassar/TelegramBotAPI/SimpleInputFilesTest.kt
@@ -13,6 +13,7 @@ private val nonstrictJsonFormat = Json {
     ignoreUnknownKeys = true
     allowSpecialFloatingPointValues = true
     useArrayPolymorphism = true
+    encodeDefaults = true
 }
 
 class SimpleInputFilesTest {