+++ TelegramBotAPI-core/src/commonTest/kotlin/com/github/insanusmokrassar/TelegramBotAPI/TestsJsonFormat.kt
@@ -2,4 +2,6 @@ package com.github.insanusmokrassar.TelegramBotAPI
 
 import kotlinx.serialization.json.Json
 
-val TestsJsonFormat = Json {}
+val TestsJsonFormat = Json {
+    encodeDefaults = true
+}