+++ TelegramBotAPI-core/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InlineQueries/InlineQueryResult/abstracts/InlineQueryResult.kt
@@ -1,12 +1,13 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult.abstracts
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult.serializers.InlineQueryResultSerializer
+import com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueryIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
 import kotlinx.serialization.Serializable
 
 @Serializable(InlineQueryResultSerializer::class)
 interface InlineQueryResult {
     val type: String
-    val id: String
+    val id: InlineQueryIdentifier
     val replyMarkup: InlineKeyboardMarkup?
 }